[cmake] Fix check for debug static build of pcre2
The correct name is `pcre2-8-staticd.lib`, not `pcre2-8d-static.lib`.

If the static debug build is the only one available, the value of
`default_pcre_libraries` would previously default to `pcre.lib`, and so
there was a linking error if this library did not exist.

Now the library will be correctly detected, which avoids errors if the
static debug build is the only one available.