Fix "Failed linking C shared library mod_md.so"
The cURL variable was declared but not assigned, resulting in a "Failed linking C shared library mod_md.so" error because cURL was not found. This change will automatically find and assign values to cURL variables, fixing the error that cURL cannot be found during HTTPD compilation without having to recompile cURL again as the solution in this error: https://bz.apache.org/bugzilla/show_bug.cgi?id=65602.