Merge r1909137, r1911067 from trunk:
mod_alias: When an alias is declared inside a Location, make sure
the balance of the URL is preserved to match the alias declared
outside a location. Fixes an error where all requests are mapped
to the root of the location.


mod_alias: Add AliasPreservePath directive to map the full
path after the alias in a location.

Submitted by: minfrin
Reviewed by: minfrin, rpluem, covener


git-svn-id: https://svn.apache.org/repos/asf/httpd/httpd/branches/2.4.x@1913009 13f79535-47bb-0310-9956-ffa450edef68