mod_proxy: Avoid AH01059 parsing error for SetHandler "unix:" URLs.  PR 69160
The hostname part of the URL is not mandated for UDS though the canon_handler
hooks will require it, so add "localhost" if it's missing (won't be used anyway
for an AF_UNIX socket).

This can trigger with SetHandler "unix:" URLs which are now also fixed up.

git-svn-id: https://svn.apache.org/repos/asf/httpd/httpd/trunk@1919015 13f79535-47bb-0310-9956-ffa450edef68
(cherry picked from commit 530106b2c1be37f581e269be30a0f2665209995e)